Delete almost works

+++ todo/app.rb
@@ -39,12 +39,22 @@ post '/done' do
   { :id => params[:id], :status => value }.to_json
 end
 
-get '/delete/:id' do
+# delete '/:id' do
+#   @items.delete(params[:id])
+#   redirect '/'
+# end
+
+get '/delete/:id/?' do
   @item = Item.first(:id => params[:id])
 end
 
-post '/delete/:id' do
-  item = Item.first(:id => params[:id])
-  item.destroy
+post '/delete/:id/?' do
+  if params.value?('OK')
+    item = Item.first(:id => params[:id])
+    item.destroy
+    puts "bitch please"
+  else
+    puts "dupa"
+  end
   redirect '/'
 end
